requirements.txt
python-telegram-bot==21.6
anthropic==0.77.0
claude-agent-sdk
python-dotenv==1.0.1
psutil==6.1.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
apscheduler==3.10.4
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
